feat: integrate treesitter-markdown
Derive our grammar from the markdown one. This requires pulling in
`treesitter-markdown` as a submodule so that we can also pull in the
external scanner.

Note that this probably breaks all the bindings which now need to
include the scanners' source code. This will be fixed in a follow-up PR.